Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PDEOptimizationProblems.meta

PDEOptimizationProblems.burger1d

PDEOptimizationProblems.cellincrease

PDEOptimizationProblems.cellincrease_MichaelisMenten

PDEOptimizationProblems.controlelasticmembrane1

PDEOptimizationProblems.controlelasticmembrane2

PDEOptimizationProblems.incompressiblenavierstokes

PDEOptimizationProblems.inversepoissonproblem2d

PDEOptimizationProblems.membrane

PDEOptimizationProblems.penalizedpoisson

PDEOptimizationProblems.poisson3d

PDEOptimizationProblems.poissonboltzman2d

PDEOptimizationProblems.smallestlaplacianeigenvalue

​

PDEOptimizationProblems.meta — Constant

PDEOptimizationProblems.meta

A composite type that represents the main features of the PDE-constrained optimization problem. optimize ∫( f(θ, y, u) )dΩ subject to lvar ≤ (θ, y, u) ≤ uvar ∫( res(θ, y, u, v) )dΩ = 0 –- The following keys are valid: Problem meta

domaindim: dimension of the domain 1/2/3 for 1D/2D/3D

pbtype: in pbtypes

nθ: size of the unknown vector

ny: number of unknown function

nu: number of control function

Solution meta

optimalvalue: best known objective value (NaN if unknown, -Inf if unbounded problem)

Classification

objtype: in objtypes

contype: in contypes

origin: in origins

has_cvx_obj: true if the problem has a convex objective

has_cvx_con: true if the problem has convex constraints

has_bounds: true if the problem has bound constraints

has_fixed_variables: true if it has fixed variables

PDEOptimizationProblems.burger1d — Method

Burger1d(;n :: Int = 512, kwargs...)

Let Ω=(0,1), we solve the one-dimensional ODE-constrained control problem: min{y,u} 0.5 ∫Ω​ |y(x) - yd(x)|^2dx + 0.5 * α * ∫Ω​ |u|^2 s.t. -ν y'' + yy' = u + h, for x ∈ Ω, y(0) = 0, y(1)=-1, for x ∈ ∂Ω, where the constraint is a 1D stationary Burger's equation over Ω, with h(x)=2(ν + x^3) and ν=0.08. The first objective measures deviation from the data y_d(x)=-x^2, while the second term regularizes the control with α = 0.01.

This example has been used in Section 9.1

The specificity of the problem:

quadratic objective function;

nonlinear constraints with AD jacobian;

Suggestions:

FEOperatorFromTerms has only one term. We might consider splitting linear and

nonlinear terms.

PDEOptimizationProblems.cellincrease — Method

Mairet, F., & Bayen, T. (2021). The promise of dawn: microalgae photoacclimation as an optimal control problem of resource allocation. Journal of Theoretical Biology, 515, 110597.

Using a photosynthetic rate proportional to the photosynthetic apparatus mass fraction.

PDEOptimizationProblems.cellincrease_MichaelisMenten — Method

Mairet, F., & Bayen, T. (2021). The promise of dawn: microalgae photoacclimation as an optimal control problem of resource allocation. Journal of Theoretical Biology, 515, 110597.

Using Michaelis-Menten's function for the photosynthetic rate.

PDEOptimizationProblems.controlelasticmembrane1 — Method

controlelasticmembrane1(; n :: Int = 10, args...)

Let Ω = (-1,1)^2, we solve the following distributed Poisson control problem with Dirichlet boundary:

min{y ∈ H^10,u ∈ H^1} 0.5 ∫Ω​ |y(x) - yd(x)|^2dx + 0.5 * α * ∫Ω​ |u|^2 s.t. -Δy = h + u, for x ∈ Ω y = 0, for x ∈ ∂Ω umin(x) <= u(x) <= umax(x) where yd(x) = -x[1]^2 and α = 1e-2. The force term is h(x1,x2) = - sin( ω x1)sin( ω x2) with ω = π - 1/8. In this first case, the bound constraints are constants with umin(x) = 0.0 and umax(x) = 1.0.

PDEOptimizationProblems.controlelasticmembrane2 — Method

controlelasticmembrane2(; n :: Int = 10, args...)

Let Ω = (-1,1)^2, we solve the following distributed Poisson control problem with Dirichlet boundary:

min{y ∈ H^10,u ∈ H^1} 0.5 ∫Ω​ |y(x) - yd(x)|^2dx + 0.5 * α * ∫Ω​ |u|^2 s.t. -Δy = h + u, for x ∈ Ω y = 0, for x ∈ ∂Ω umin(x) <= u(x) <= umax(x) where yd(x) = -x[1]^2 and α = 1e-2. The force term is h(x1,x2) = - sin( ω x1)sin( ω x2) with ω = π - 1/8. In this second case, the bound constraints are umin(x) = x1+x2 and umax(x) = x1^2+x2^2 applied at the midpoint of the cells.

PDEOptimizationProblems.incompressiblenavierstokes — Method

incompressibleNavierStokes(; n :: Int64 = 3, kargs...)

This corresponds to the incompressible Navier-Stokes equation described in the Gridap Tutorials: https://gridap.github.io/Tutorials/stable/pages/t008incnavier_stokes/

It has no objective function and no control, just the PDE.

PDEOptimizationProblems.inversepoissonproblem2d — Method

inversePoissonproblem2d(;n :: Int = 512, kwargs...)

Let Ω=(-1,1)^2, we solve the 2-dimensional PDE-constrained control problem: min{y ∈ H1^0, u ∈ L^∞} 0.5 ∫Ω​ |y(x) - yd(x)|^2dx + 0.5 * α * ∫Ω​ |u|^2 s.t. -∇⋅(z∇u) = h, for x ∈ Ω, u(x) = 0, for x ∈ ∂Ω. Let c = (0.2,0.2) and and define S1 = {x | ||x-c||2 ≤ 0.3 } and S2 = {x | ||x-c||1 ≤ 0.6 }. The target ud is generated as the solution of the PDE with z*(x) = 1 + 0.5 * I{S1}(x) + 0.5 * I{S2}(x). The force term here is h(x1,x2) = - sin( ω x1)sin( ω x2) with ω = π - 1/8. The control variable z represents the diffusion coefficients for the Poisson problem that we are trying to recover. Set α = 10^{-4} and discretize using P1 finite elements on a uniform mesh of 1089 triangles and employ an identical discretization for the optimization variables u, thus ncon = 1089 and npde = 961. Initial point is y0=1 and u_0 = 1. z ≥ 0 (implicit)

This example has been used in Section 9.2

PDEOptimizationProblems.membrane — Method

https://arxiv.org/pdf/2103.14552.pdf Example 1. MEMBRANE Multilevel Active-Set Trust-Region (MASTR) Method for Bound Constrained Minimization Alena Kopaničáková and Rolf Krause

The solution and original problem is given in Domorádová, M., & Dostál, Z. (2007). Projector preconditioning for partially bound‐constrained quadratic optimization. Numerical Linear Algebra with Applications, 14(10), 791-806.

PDEOptimizationProblems.penalizedpoisson — Method

Let Ω=(0,1)^2, we solve the unconstrained optimization problem: min{u ∈ H1^0} 0.5 ∫_Ω​ |∇u|^2 - w u dx s.t. u(x) = 0, for x ∈ ∂Ω whre w(x)=1.0.

The minimizer of this problem is the solution of the Poisson equation: ∫_Ω​ (∇u ∇v - f*v)dx = 0, ∀ v ∈ Ω u = 0, x ∈ ∂Ω

This example has been used in Exercice 10.2.4 (p. 308)

PDEOptimizationProblems.poisson3d — Method

poisson3d(; n :: Int = 10)

This example represents a Poisson equation with Dirichlet boundary conditions over the 3d-box, (0,1)^3, and we minimize the squared H_1-norm to a manufactured solution. So, the minimal value is expected to be 0.

It is inspired from the 2nd tutorial in Gridap.jl: https://gridap.github.io/Tutorials/stable/pages/t002_validation/

PDEOptimizationProblems.poissonboltzman2d — Method

poissonBoltzman2d(; n :: Int = 100)

Let Ω=(-1,1)^2, we solve the 2-dimensional PDE-constrained control problem: min{y ∈ H1^0, u ∈ L^∞} 0.5 ∫Ω​ |y(x) - yd(x)|^2dx + 0.5 * α * ∫Ω​ |u|^2 s.t. -Δy + sinh(y) = h + u, for x ∈ Ω y(x) = 0, for x ∈ ∂Ω

The force term here is h(x1,x2) = - sin( ω x1)sin( ω x2) with ω = π - 1/8. The targeted function is yd(x) = {10 if x ∈ [0.25,0.75]^2, 5 otherwise}. We discretize using P1 finite elements on a uniform mesh with 10201 triangles, resulting in a problem with n = 20002 variables and m = 9801 constraints. We use y0=1 and u0 = 1 as the initial point.

This example has been used in Section 9.3

The specificity of the problem:

quadratic objective function;

nonlinear constraints with AD jacobian;

PDEOptimizationProblems.smallestlaplacianeigenvalue — Method

smallestLaplacianeigenvalue(; n :: Int = 10, args...)

We solve the following problem:

min{u,z} ∫Ω​ |∇u|^2 s.t. ∫_Ω​ u^2 = 1, for x ∈ Ω u = 0, for x ∈ ∂Ω

The solution is an eigenvector of the smallest eigenvalue of the Laplacian operator, given by the value of the objective function. λ is an eigenvalue of the Laplacian if there exists u such that

Δu + λ u = 0, for x ∈ Ω u = 0, for x ∈ ∂Ω

This example has been used in Exercice 10.2.11 (p. 313) and more eigenvalue problems can be found in Section 7.3.2

TODO:

does the 1 work as it is? or should it be put in lcon, ucon?

it is 1D for now.